requirements.txt
pandas>=2.0.0
geopandas>=0.14.0
requests>=2.31.0
overpy>=0.7
sqlalchemy>=2.0.0
arcgis>=2.3.0
shapely>=2.0.0
pyproj>=3.6.0
fiona>=1.9.0
numpy>=1.24.0
python-dotenv>=1.0.0
tqdm>=4.65.0
